Review
Hi Sophia, 

I made a few minor revisions to your course. These are so that the course follows basic conventions for mathematical explanations. A few of these conventions are:

* Use "we" instead of "I" or "you". The reasoning behind this is to make the reader feel like they are not being lectured by some person, but rather reading information at their own free will. Most users prefer this language. For examples, see edits on lines 107, 111, and 113.
* Align math to center and use display style if the expression is an equality, uses complex figures, and/or exceeds 3 terms. (110, 112, 114)
* When writing question papers, separate math from english as much as possible. (120 => 118, 121)

What you've been writing is amazing as it follows all the other standards, such as referencing prior and general knowledge to build on it, clearly stating new terms, writing solutions without assuming any knowledge (I need to work on this!) and overall making sure there is no confusion.

One other thing you might want to work on is doing a one or two sentence recap in the start of each lesson, describing what the reader learned last lesson. Additionally, the same goes for the end of the lesson where you let them know what to look forward to and acknowledge any pressing questions they might have. The integrals course does a good job of this. This one is not so important though, and sometimes it does not fit into the context and we have to let it go. However, the recap is quite important.